renaming interface
