+++ vivid-server/src/models/guild.interface.ts
