+++ vivid-server/src/modules/guilds/guilds.controller.ts
@@ -16,7 +16,7 @@ import { GuildsEntity } from '@/guilds.entity';
 import { UserEntity } from '@/user.entity';
 import { UserService } from '$/users/user.service';
 import { GuildsService } from './guilds.service';
-import { IGuild } from '@/user.interface copy';
+import { IGuild } from '~/models/guild.interface';
 
 @Controller('guilds')
 @UseGuards(AuthenticatedGuard)
